+++ app/CHANGELOG.md
@@ -1,4 +1,6 @@
 #####0.13####
+first public release!
+
 changes:
 
  - users can select the protocol (HLS/RTSP/RTMP) to open streams up locally. casting a stream will still be HLS regardless of setting